Refactoring based on Feedback from July 26
turned IBOutlet properties into weak references

changed int --> NSInteger, float --> CGFloat, long --> NSInteger

moved static declarations to be above @implementation (except for static singleton tokens in initializers)

fixed formatting error + refactored while loop in Event.m to be a fast enumeration

+++ fbu-project/fbu-project/Utility/UserSorter.h
@@ -15,10 +15,10 @@ extern NSArray *_Nullable Recommended(void);
 
 extern NSArray<PFUser *> *_Nullable NotLikedUsers(void);
 extern CLLocation * LocationForUser(PFUser *_Nonnull user);
-extern double DistanceBetweenUsers(PFUser *_Nonnull user1, PFUser *_Nonnull user2);
-extern int PointsForDistance(double distance);
+extern CGFloat DistanceBetweenUsers(PFUser *_Nonnull user1, PFUser *_Nonnull user2);
+extern NSInteger PointsForDistance(double distance);
 extern NSSet<NSString *> * SetWithGenreTitles(NSArray<LikedGenre *> *_Nonnull likedGenres);
-extern int PointsForCommonGenres(PFUser *_Nonnull user1, PFUser *_Nonnull user2);
+extern NSInteger PointsForCommonGenres(PFUser *_Nonnull user1, PFUser *_Nonnull user2);
 NSInteger PointsForMatchingInstrumentsInRecording(PFUser *_Nonnull otherUser);
 extern NSNumber * RankUserForUser(PFUser *_Nonnull recommendedUser, PFUser *_Nonnull currentUser);
